backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.1",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.3",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "cachecontrol==0.14.3",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "colorama==0.4.6",
    "cryptography==46.0.3",
    "fastapi==0.120.0",
    "firebase-admin==6.5.0",
    "frozenlist==1.8.0",
    "google-ai-generativelanguage==0.6.6",
    "google-api-core>=2.25.2",
    "google-api-python-client==2.185.0",
    "google-auth==2.41.1",
    "google-auth-httplib2==0.2.0",
    "google-cloud-core==2.4.3",
    "google-cloud-firestore==2.21.0",
    "google-cloud-storage==3.4.1",
    "google-crc32c==1.7.1",
    "google-genai",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.71.0",
    "grpcio==1.76.0",
    "grpcio-status==1.62.3",
    "h11==0.16.0",
    "httplib2==0.31.0",
    "idna==3.11",
    "msgpack==1.1.2",
    "multidict==6.7.0",
    "propcache==0.4.1",
    "proto-plus==1.26.1",
    "protobuf==4.25.8",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pyjwt==2.10.1",
    "pyparsing==3.2.5",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "starlette==0.48.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "yarl==1.22.0",
]
